Time: 3 ms (18.47%), Space: 48.4 MB (45.56%) - LeetHub

+++ 3685-count-subarrays-of-length-three-with-a-condition/3685-count-subarrays-of-length-three-with-a-condition.cpp
@@ -8,7 +8,7 @@ public:
             int j = i + 1;
             int k = i + 2;
             if (j < n && k < n) {
-                if ((double)nums[i] + nums[k] == (double)nums[j] / 2) {
+                if (nums[i] + nums[k] == nums[j] / 2.0) {
                     count++;
                 }
             }
